cli: make set of immutable commits configurable
This adds a new `revsets.immutable-heads` config for defining the set
of immutable commits. The set is defined as the configured revset, as
well as its ancestors, and the root commit commit (even if the
configured set is empty).

This patch also adds enforcement of the config where we already had
checks preventing rewrite of the root commit. The working-copy commit
is implicitly assumed to be writable in most cases. Specifically, we
won't prevent amending the working copy even if the user includes it
in the config but we do prevent `jj edit @` in that case. That seems
good enough to me. Maybe we should emit a warning when the working
copy is in the set of immutable commits.

Maybe we should add support for something more like [Mercurial's
phases](https://wiki.mercurial-scm.org/Phases), which is propagated on
push and pull. There's already some affordance for that in the view
object's `public_heads` field. However, this is simpler, especially
since we can't propagate the phase to Git remotes, and seems like a
good start. Also, it lets you say that commits authored by other users
are immutable, for example.

For now, the functionality is in the CLI library. I'm not sure if we
want to move it into the library crate. I'm leaning towards letting
libraries do whatever they want without being restricted by immutable
commits. I do think we should move the functionality into a future
`ui-lib` or `ui-util` crate. That crate would have most of the
functionality in the current `cli_util` module (but in a
non-CLI-specific form).

I called the config `immutable-heads`, but I'm not sure if that's
better than just `immutable`. My main reasons for calling it that is
so it's clear that they don't need to include ancestors and to make it
not confusingly similar to the planned `immutable()` revset (which
will include ancestors and the root commit).